creando archivos para conectar con firebase, creamos en la consola de firebase la base de datos y subimos las aplicaciones tanto de android como de Ios con sus correspondientes identificadores, luego finalizamos de construir el file de services.dart y adicional cargamos en cada carpeta los archivos dados al crear la app en firebase que serian googleServices-Info para Ios en la carpeta Runner y google-services.js para android en la android sub carpeta app

+++ lib/book_details/book_details.dart
@@ -33,7 +33,7 @@ class BookdetailScreen extends StatelessWidget {
             BookInfoDetailsWidget(_book.title, _book.author, _book.description),
             Padding(
               padding: const EdgeInsets.only(top: 45),
-              child: ButtonActionBookWidget(_book.id),
+              child: ButtonActionBookWidget(_book.id as int),
             ),
           ],
         ),
